% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{categorical_vars}
\alias{categorical_vars}
\title{categorical variables}
\usage{
categorical_vars(df)
}
\arguments{
\item{df}{dataframe to be evaluated}
}
\value{
logical
}
\description{
Determine all categorical variables of a dataframe and return a vector of their names.
}
\examples{
categorical_vars(CRC305ABC_DM)
}
