% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central.R
\name{domain_ids}
\alias{domain_ids}
\title{domain identifier variables}
\usage{
domain_ids(df)
}
\arguments{
\item{df}{dataframe of one SDTM domain}
}
\value{
character vector of identifier variable names
}
\description{
Return identifier variables of a domain.
}
\examples{
domain_ids(CRC305ABC_DM)
}
