% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Hk}
\alias{Hk}
\title{Shannon entropy of subsets of dataframe variables}
\usage{
Hk(.data, k)
}
\arguments{
\item{.data}{dataframe}

\item{k}{integer indicating the cardinality of the subsets of \code{.data}'s variables}
}
\value{
Numeric vector of Shannon entropy of variable subset.
  Return value names are composed by pasting together variable names using "~" as separator.
}
\description{
For all subsets of dataframe variables of a certain cardinality compute the 
Shannon entropy.
}
\examples{
\dontrun{Hk(CRC305ABC_DM, 1)}
}
\keyword{internal}
