% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Hmax}
\alias{Hmax}
\title{maximum entropy}
\usage{
Hmax(N)
}
\arguments{
\item{N}{a number, a dataframe, or a matrix}
}
\value{
log(N, base=2)
}
\description{
Compute maximum shannon entropy \code{log(N, base=2)} of a number or of a dataframe or matrix. 
In the latter case the number of rows is used.
}
\examples{
Hmax(CRC305ABC_DM)
}
