% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Hstruc}
\alias{Hstruc}
\title{Shannon entropy structure details of a dataframe}
\usage{
Hstruc(.data)
}
\arguments{
\item{.data}{dataframe}
}
\value{
A list
}
\description{
\code{Hstruc} computes the Shannon entropy of each dataframe's variable and
checks for \itemize{
 \item constant,
 \item record identifying, or
 \item 1:1 equivalent
 }
variables.
}
\examples{
Hstruc(CRC305ABC_DM)
}
