% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Hvars}
\alias{Hvars}
\title{joint Shannon entropy}
\usage{
Hvars(.data, ...)
}
\arguments{
\item{.data}{dataframe within which entropy is computed}

\item{...}{variables of the dataframe. Can be unquoted comma separated variables
or a character vector of variable names \code{all_of(vars)}}
}
\value{
numeric value of joint Shannon entropy
}
\description{
joint Shannon entropy
}
\examples{
\dontrun{Hvars(CRC305ABC_DM, SEX, RACE)

vars <- c('SEX', 'RACE')
Hvars(CRC305ABC_DM, all_of(vars))}
}
\keyword{internal}
