% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central.R
\name{synth_df}
\alias{synth_df}
\title{generate synthesis dataframe}
\usage{
synth_df(USUBJID, n_new = 10, width = 3, maxweight = 2/3)
}
\arguments{
\item{USUBJID}{character vector of person identifiers}

\item{n_new}{number of new persons to generate, defaults to 10}

\item{width}{number of persons from original trial to use for new person synthesis, defaults to 3}

\item{maxweight}{maximum allowed weight for one person for synthesis of new person, defaults to 2/3}
}
\value{
The synthesis tibble relates new synthetic persons to source persons and 
  specifies the weight of each source person's contribution.
}
\description{
\code{synth_df()} generates the tibble that relates original and synthetic new persons.
This synthesis dataframe should be kept secret to protect privacy of the original persons;
it is the base for all synthetic dataframes/tibbles that are generated.
}
\examples{
synth_df(USUBJID = letters)
}
