% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central.R
\name{synthesize}
\alias{synthesize}
\title{synthesize a SDTM dataframe}
\usage{
synthesize(syndf, df, cat_fuzz = 1)
}
\arguments{
\item{syndf}{synthesis dataframe}

\item{df}{original SDTM dataframe}

\item{cat_fuzz}{fuzz factor for noise on categorical variables, defaults to 1}
}
\value{
synthesized SDTM dataframe
}
\description{
Create a synthetic dataframe from syndf and a source SDTM dataframe.
}
\examples{
synthesize(synth_df(CRC305ABC_DM$USUBJID), CRC305ABC_DM)
}
