% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central.R
\name{synthesize_E}
\alias{synthesize_E}
\title{synthesize events dataframe}
\usage{
synthesize_E(syndf, df, cat_fuzz = 1)
}
\arguments{
\item{syndf}{synthesis dataframe}

\item{df}{original SDTM dataframe}

\item{cat_fuzz}{fuzz factor for noise on categorical variables, defaults to 1}
}
\value{
synthesized SDTM events dataframe
}
\description{
synthesize events dataframe
}
\examples{
\dontrun{synthesize_E(synth_df(CRC305ABC_DM$USUBJID), CRC305ABC_AE)}
}
