% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntrial.R
\docType{package}
\name{syntrial}
\alias{syntrial}
\title{syntrial: A package for generating synthetic twin clinical trials from a real one.}
\description{
The syntrial package provides three categories of important functions:
syntrial, bar and baz.
}
\section{syntrial functions}{

The syntrial functions ...
}

